requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
colorama==0.3.9
coreapi==2.3.3
coreapi-cli==1.0.6
coreschema==0.0.4
coverage==4.3.4
decorator==4.3.0
defusedxml==0.5.0
Django==2.0.6
django-allauth==0.36.0
django-braces==1.13.0
django-debug-toolbar==1.9.1
django-extensions==2.0.7
django-filter==2.0.0
django-ranged-response==0.2.0
django-rest-swagger==2.2.0
django-simple-captcha==0.5.9
djangorestframework==3.8.2
idna==2.7
ipdb==0.10.2
ipython==5.3.0
ipython-genutils==0.2.0
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
oauthlib==2.1.0
openapi-codec==1.3.2
pickleshare==0.7.4
Pillow==5.2.0
prompt-toolkit==1.0.15
psycopg2-binary==2.7.4
Pygments==2.2.0
python3-openid==3.1.0
pytz==2018.4
requests==2.19.1
requests-oauthlib==1.0.0
simplegeneric==0.8.1
simplejson==3.16.0
six==1.11.0
sqlparse==0.2.4
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.23
wcwidth==0.1.7
